Improve out-of-band signaling that network data is ready.
* Remove rxBytesWaiting from NetworkStatus class since it's never current
* Remove timer and assert_interrupt() from drivewire network drive
* Change poll_interrupt() to return true/false if there's new data

+++ lib/network-protocol/SSH.h
@@ -134,7 +134,7 @@ private:
      * Return if bytes available by injecting into RX buffer.
      * @return number of bytes available
      */
-    unsigned short available();
+    size_t available();
 };
 
 #endif /* NETWORKPROTOCOL_SSH */
